adding a refactored test case
